Re-initialize zone after each refresh

+++ indisoluble/a_healthy_dns/tools/can_create_connection.py
@@ -4,7 +4,7 @@ import logging
 import socket
 
 
-def can_create_connection(ip: str, port: int, timeout: int) -> bool:
+def can_create_connection(ip: str, port: int, timeout: float) -> bool:
     try:
         with socket.create_connection((ip, port), timeout):
             logging.debug("TCP connectivity test to '%s:%d' successful", ip, port)
